feat(dashboard): create a custom hook usePaginatedRecords for handling pagination, implement paginated records hook
